rename SpectrumView -> SpectrumView_v1
