+++ AstraBox/Views/SpectrumView_v1.py
